src/Learnify.Subtitles/requirements.txt
pika
python-dotenv
pydub
git+https://github.com/openai/whisper.git
azure-storage-blob